Review comment: Removed useless double line check
That duplicate line was there since when the code was imported. At first
I though that this possible used references to change the coordinates,
but that was not the case, so removing it should be safe.
Removing it did not cause any problems with lines in the power system
test scenario.