GameObject: Refactor names in CrossCheck
Also reordered the variable initialization. Not sure why the values are
|=-assigned in several lines, instead of doing that when the variable is
initialized.

The variable names *may* be confusing. I tried to find suitable names,
such as "active/passive", "offender/defender", "hammer/anvil", but I
stuck with "ball/goal", because it fits both the collision and the
collection purpose of the check really well.
In case that it is too abstract we can rename it to "target/projectile",
of course, as some of the comments in that function already suggest.