Refactor shape (#91)
Refactor Shape

* Code formatting: Brackets, whitespace, comments, variable definitions in separate lines, initialize variables the first time they are used
* Removed duplicate if-block:
That duplicate line was there since when the code was imported. At first I though that this possibly used references to change the coordinates, but that was not the case, so removing it should be safe.
Removing it did not cause any problems with lines in the power system test scenario.