script GUI: make sure windows can be closed only once (closes GH#104)
Fixes crash when a window has multiple children with a child with an object as its target and the object is removed and the closing callback removes the object.
E.g. on game cleanup, all objects would be cleared, closing their respective menu windows which would call the closing callbacks. The children of the child could then be deleted while the child still iterated over them.

The previous check just made sure that the closing callback is not called twice. This is not enough, though.