script GUI: prevent deleting objects from related parts of the window tree from a window closing callback (related to GH#104)
Previously, the following structure would lead to a crash {{Target = obj2, {Target = obj1, OnClose = removeobject2}, {Target = obj1}}, {}, {}}; RemoveObject(obj1);
This solution is not ideal as it potentially can (not sure!) leave deleted Target pointers somewhere in the tree or leak memory. This is still better than crashing without a hint.