Fix bug #2048: Duplicate removal by AbortCall
Assume you have an object with action "AbortAction" that has an
AbortCall="RemoveObject", or an abort call that removes the object.

Then, removing the object while it has "AbortAction" will remove it
twice, reducing the Def->Count by 2, while only one object was removed.

This does not do very much, but if you have only one object of that type
and later check for object count after you create a new one, then
creating the new object increases the Def->Count from -1 to 0 and the
new object is not found by Find_ID(), even though it exists.