Removed player number from the engine
Without the player start index from Scenario.txt there is no real need for it anymore. Replaced it with player ID, which is unique.
Unfortunately, the player ID starts at 1 now, which is inconvenient for saving player stats in arrays etc., so in the next commit this will be fixed.
Documentation will be updated later.