Fix deprecation warnings in macOS.
This replaces a number of deprecated enums and bitmasks with their renames.

It replaces the call to Gestalt (deprecated years ago) with a check
against NSAppKitVersionNumber.

It also defines GL_SILENCE_DEPRECATION, which silences the warnings
about OpenGL being deprecated on macOS (we're not going to switch to
Metal yet).