Fix build of netpuncher and tests on macOS, and add a Travis macOS build.
StdSchedulerMac needs to be included in libmisc so that the netpuncher
and tests targets can pick it up.  This is conditional on APPLE being set
(unlike the other StdScheduler* files) because the Travis builds for other
OSes do not have the Objective C++ compiler installed.  Also, there is an
earlier cset that implies that this was a problem more generally (515aaf6).

Clang/libc++ on macOS apparently doesn't support quick_exit, even with
-std=c++11.  Use a macro check in AulDeathTest to switch it to regular
exit instead.

Add test build output and autogen directories to .gitignore.