Fix slowness in deleting proplist references
// Example script
global func Main() {
    var x = Bla(1000 * 12);
    Log("Done");
}

global func Bla(count) {
    if (count < 1)
        return nil;
    count -= 1;
    var x = new Global { i = count };
    var a = Bla(count / 2);
    var b = Bla(count - count / 2);
    x.b = b;
    x.a = a;
    return x;
}

The phenomenon occurs depending on whether the order of creating a and b
is the same as s assigning them to x or not. On some architectures/
compiler combinations, the problem may also occur randomly, regardless
of the order.

Execution times vary between 15 ms and 700 ms without this patch.