SDL: Cache available multisampling settings
The method we use to enumerate the available settings with SDL sucks and
breaks the game on Wayland when used after the main window was
initialized. Determining this once on startup and caching the result
kind of fixes the issue, but startup still seems to be slowed down by
this.